knowledge/source-packets/ams-io-bench-and-ams-io-agent-2026.aaai/official-code/requirements.txt
# AI Assistant Core Dependencies
smolagents[toolkit]
smolagents[openai]

# Environment Variable Management
python-dotenv

# Configuration Management
pyyaml

# Virtuoso Environment Interaction
skillbridge

# Optional: Web Interface Support
smolagents[gradio]

# Data Processing
openpyxl

# HTTP Requests (for image vision API)
requests>=2.28

# Visualization
matplotlib

# Testing
pytest